Move the XLA tensor methods implementing the ATEN function into a new tensor_methods.cpp file. Re-organized the order of the ATEN mapped APIs in alphabetical order.
